Update README.md to include generic integral
The procedures outlined above should help users set up and run the generic integral function test case. I have verified that those instructions will allow the user to set up and execute the run, as long as they have compiled the executable. 

This update is only for the documentation. There is also a `prepare_run` script that could be updated, but perhaps that should be a separate issue and pull request. 

I hope this is okay - this is my first attempt at a pull request! Please let me know if I need to change anything.